backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.28.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
pgvector==0.2.4
sentence-transformers==2.5.1
httpx==0.26.0
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
apscheduler==3.10.4
